chat extract form to its own component
