+++ apps/thundermoon/test/thundermoon/chat/chat_messages_test.exs
@@ -8,12 +8,12 @@ defmodule Thundermoon.ChatMessagesTest do
   end
 
   test "add message" do
-    ChatMessages.add(%{user: "crumb", text: "fritz the cat"})
-    assert ChatMessages.list() == [%{user: "crumb", text: "fritz the cat"}]
+    ChatMessages.add(%{user: "crumb", text: "fritz the cat", user_id: 5})
+    assert ChatMessages.list() == [%{user: "crumb", text: "fritz the cat", user_id: 5}]
   end
 
   test "clear all messages" do
-    ChatMessages.add(%{user: "crumb", text: "fritz the cat"})
+    ChatMessages.add(%{user: "crumb", text: "fritz the cat", user_id: 5})
     ChatMessages.clear()
     assert ChatMessages.list() == []
   end
